image to text
created image_to_text.dart which can be imported and called by either create card or edit card.

should allow for image to text from a filepath or camera image.

I think permissions have been set for both Ios and Android.

MacOs requires permissions for using the camera, but does not for selecting from a file. Using the camera on the computer seems clunky, so during implementation, perhaps we limit to selecting from a file depending on the platform for either windows or macOs.

have not run a test, but currently has no nastigrams, and based on tutorials i've found will hopefully work smoothly