camera_screen functionality
This should make camera screen functional. The one thing I can't figure out is the stupid textField to display the scanned text. When I add it, everything stops rendering. For right now I have that portion commented out. if someone fixes this, this page is nearly ready to make cards