card and camera screen cleanup
I made a mess last round with flashcard state global keys rather than creating an instance of card because programming on no sleep in a language you don't know well is baaaadddd.....   this cleans up the mess I made last time. it still isn't working. in card preview when you click the add to front button any typed text will get added to the front bullet points array but only if selected. it however does not update the appearance on the screen. the debugPrint shows the add button is working